Add edge_groups.{cpp,h}, pass context into set_device_context()
